+++ crates/ruma-client-api/src/error.rs
@@ -27,7 +27,11 @@ mod kind_serde;
 #[non_exhaustive]
 pub enum ErrorKind {
     /// M_FORBIDDEN
-    Forbidden,
+    Forbidden {
+        /// The `WWW-Authenticate` header error message.
+        #[cfg(feature = "unstable-msc2967")]
+        authenticate: Option<AuthenticateError>,
+    },
 
     /// M_UNKNOWN_TOKEN
     UnknownToken {
@@ -199,7 +203,7 @@ pub struct Extra(BTreeMap<String, JsonValue>);
 impl AsRef<str> for ErrorKind {
     fn as_ref(&self) -> &str {
         match self {
-            Self::Forbidden => "M_FORBIDDEN",
+            Self::Forbidden { .. } => "M_FORBIDDEN",
             Self::UnknownToken { .. } => "M_UNKNOWN_TOKEN",
             Self::MissingToken => "M_MISSING_TOKEN",
             Self::BadJson => "M_BAD_JSON",
@@ -303,10 +307,6 @@ pub struct Error {
     /// The http status code.
     pub status_code: http::StatusCode,
 
-    /// The `WWW-Authenticate` header error message.
-    #[cfg(feature = "unstable-msc2967")]
-    pub authenticate: Option<AuthenticateError>,
-
     /// The http response's body.
     pub body: ErrorBody,
 }
@@ -314,12 +314,7 @@ pub struct Error {
 impl Error {
     /// Constructs a new `Error` with the given status code and body.
     pub fn new(status_code: http::StatusCode, body: ErrorBody) -> Self {
-        Self {
-            status_code,
-            #[cfg(feature = "unstable-msc2967")]
-            authenticate: None,
-            body,
-        }
+        Self { status_code, body }
     }
 
     /// If `self` is a server error in the `errcode` + `error` format expected
@@ -333,16 +328,24 @@ impl EndpointError for Error {
     fn from_http_response<T: AsRef<[u8]>>(response: http::Response<T>) -> Self {
         let status = response.status();
 
-        #[cfg(feature = "unstable-msc2967")]
-        let authenticate = response
-            .headers()
-            .get(http::header::WWW_AUTHENTICATE)
-            .and_then(|val| val.to_str().ok())
-            .and_then(AuthenticateError::from_str);
-
         let body_bytes = &response.body().as_ref();
         let error_body: ErrorBody = match from_json_slice(body_bytes) {
-            Ok(StandardErrorBody { kind, message }) => ErrorBody::Standard { kind, message },
+            Ok(StandardErrorBody { kind, message }) => {
+                #[cfg(feature = "unstable-msc2967")]
+                let kind = if let ErrorKind::Forbidden { .. } = kind {
+                    let authenticate = response
+                        .headers()
+                        .get(http::header::WWW_AUTHENTICATE)
+                        .and_then(|val| val.to_str().ok())
+                        .and_then(AuthenticateError::from_str);
+
+                    ErrorKind::Forbidden { authenticate }
+                } else {
+                    kind
+                };
+
+                ErrorBody::Standard { kind, message }
+            }
             Err(_) => match MatrixErrorBody::from_bytes(body_bytes) {
                 MatrixErrorBody::Json(json) => ErrorBody::Json(json),
                 MatrixErrorBody::NotJson { bytes, deserialization_error, .. } => {
@@ -351,13 +354,7 @@ impl EndpointError for Error {
             },
         };
 
-        let error = error_body.into_error(status);
-
-        #[cfg(not(feature = "unstable-msc2967"))]
-        return error;
-
-        #[cfg(feature = "unstable-msc2967")]
-        Self { authenticate, ..error }
+        error_body.into_error(status)
     }
 }
 
@@ -379,12 +376,7 @@ impl std::error::Error for Error {}
 impl ErrorBody {
     /// Convert the ErrorBody into an Error by adding the http status code.
     pub fn into_error(self, status_code: http::StatusCode) -> Error {
-        Error {
-            status_code,
-            #[cfg(feature = "unstable-msc2967")]
-            authenticate: None,
-            body: self,
-        }
+        Error { status_code, body: self }
     }
 }
 
@@ -397,7 +389,11 @@ impl OutgoingResponse for Error {
             .status(self.status_code);
 
         #[cfg(feature = "unstable-msc2967")]
-        let builder = if let Some(auth_error) = &self.authenticate {
+        let builder = if let ErrorBody::Standard {
+            kind: ErrorKind::Forbidden { authenticate: Some(auth_error) },
+            ..
+        } = &self.body
+        {
             builder.header(http::header::WWW_AUTHENTICATE, auth_error)
         } else {
             builder
@@ -542,7 +538,13 @@ mod tests {
         }))
         .unwrap();
 
-        assert_eq!(deserialized.kind, ErrorKind::Forbidden);
+        assert_eq!(
+            deserialized.kind,
+            ErrorKind::Forbidden {
+                #[cfg(feature = "unstable-msc2967")]
+                authenticate: None
+            }
+        );
         assert_eq!(deserialized.message, "You are not authorized to ban users in this room.");
     }
 
@@ -613,9 +615,9 @@ mod tests {
 
         assert_eq!(error.status_code, http::StatusCode::UNAUTHORIZED);
         assert_matches!(error.body, ErrorBody::Standard { kind, message });
-        assert_eq!(kind, ErrorKind::Forbidden);
+        assert_matches!(kind, ErrorKind::Forbidden { authenticate });
         assert_eq!(message, "Insufficient privilege");
-        assert_matches!(error.authenticate, Some(AuthenticateError::InsufficientScope { scope }));
+        assert_matches!(authenticate, Some(AuthenticateError::InsufficientScope { scope }));
         assert_eq!(scope, "something_privileged");
     }
 }